fix: also apply the Splunk-JSON UTC offset in eid-metrics/logon-summary times (#1820)
Codex review of #1897 flagged that the same offset-discarding pattern the
log-metrics fix addressed still existed in src/timeline/metrics.rs
(EventMetrics::stats_time_cnt and parse_evtx_datetime), so eid-metrics /
logon-summary time ranges and the general first/last timestamps were still
skewed for Splunk-JSON input (a "+09:00" time stored as if it were UTC).

Extract the offset-aware parser into a shared `utils::parse_evtx_timestamp` and
use it from all three call sites (log_metrics + the two in metrics.rs) so the fix
cannot drift out of sync again. The evtx UTC ("...Z") path is unchanged, so
detection and metrics output is byte-identical on evtx input (verified on the
sample-evtx corpus: every timeline/metrics output identical, eid/logon
content-identical modulo the known HashMap-seed row order); only Splunk-JSON
offset timestamps change. The parser unit test moved to utils.rs.
