+++ CHANGELOG-Japanese.md
@@ -12,7 +12,7 @@
 
 - `read_jsonl_to_value`/`read_json_to_value` のファイルオープンエラーが、ファイルパスの代わりにプレースホルダー `{path}` をそのまま出力していた問題を修正した（エラー文字列が `format!` ではなく通常の文字列リテラルだった）。 (#1897) (@YamatoSecurity)
 - `eid-metrics` テーブルの「Event」列の幅計算で、55桁未満のターミナルで `u16` のアンダーフローが発生する問題を修正した。`terminal_width - 55` が45文字の下限を適用する前にアンダーフローし、オーバーフローチェック有効のビルドではパニックし、リリースビルドでは巨大な値にラップしていた（列の上限が実質無効になっていた）。飽和減算を使うようにした。 (#1897) (@YamatoSecurity)
-- `log-metrics` が Splunk-JSON 入力の First/Last Timestamp 列をずらしていた問題を修正した。タイムスタンプのフォールバックが明示的な UTC オフセット（例: `+09:00`）を `NaiveDateTime` で解析していたためオフセットが破棄され、ローカルの時計時刻がそのまま UTC として保存されていた。オフセットを適用するようにした。 (#1897) (@YamatoSecurity)
+- 明示的な UTC オフセット（例: `+09:00`）を持つ Splunk-JSON のタイムスタンプが `NaiveDateTime` で解析されてオフセットが破棄され、ローカルの時計時刻がそのまま UTC として保存されていた問題を修正した。これにより `log-metrics` の First/Last Timestamp 列に加えて `eid-metrics`/`logon-summary` の時刻範囲（`EventMetrics::stats_time_cnt` と `parse_evtx_datetime`）もずれていた。タイムラインの集計処理は、オフセットを適用する共通のパーサー（`utils::parse_evtx_timestamp`）を使うようにした。 (#1897) (@YamatoSecurity)
 - カスタムのルール設定ディレクトリ（`-c`）を指定した `-G`（GeoIP エンリッチメント）が、そのディレクトリの `geoip_field_mapping.yaml` を無視する問題を修正した。ルックアップが拡張子なしの `geoip_field_mapping` を使っていたため実ファイルに一致せず、常にデフォルトのバンドルされたマッピングにフォールバックしていた。 (#1897) (@YamatoSecurity)
 - `count(field)` 相関ルールがルールの `timeframe` より長い範囲にまたがる誤検知アラートを生成する問題を修正した。`judge_timeframe` のスライディングウィンドウで、条件を満たさなかったウィンドウを処理する分岐が左端を外す前にチェックなしの `add_data(right)` を行っていたため、ウィンドウの左端からタイムフレーム外であるとすでに判明している `records[right]` を次のウィンドウに取り込んでいた。そのレコードのフィールド値が新規の場合、タイムフレームより長い範囲でユニーク値のカウントがしきい値を超えることがあった。スライド時は左端を外すだけにして、内側のループが `right` をタイムフレームチェック付きで再度伸ばすようにしたため、タイムフレーム外のレコードはカウントされない。フィールド指定なしの `count()` は影響を受けない。 (#1896) (@YamatoSecurity)
 - 結果サマリーの狭いターミナルでの表示に関する問題を修正した: `_print_timeline_hist` が `usize` の減算をチェックせずに行っていたため、極端に狭いターミナルでアンダーフローして巨大なメモリ確保やハングを引き起こす可能性があった（描画スペースが無い場合はヒストグラムをスキップし、`saturating_sub` を使うようにした）。ルール作者テーブルの行数を実際の列数ではなくハードコードされた `is_multiple_of(4)` で計算していたため多くの幅で不揃いなテーブルになっていた（`authors_num.div_ceil(table_column_num)` に変更）。また、ルール作者テーブルが非常に狭いターミナル（72文字以下）ではみ出す/折り返す代わりに1列にフォールバックするようにした。タイムライン出力には影響しない。 (#1858) (@YamatoSecurity)
