+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 - Fixed the `read_jsonl_to_value`/`read_json_to_value` file-open error printing the literal placeholder `{path}` instead of the file path (the error string was a plain string literal rather than a `format!`). (#1897) (@YamatoSecurity)
 - Fixed a `u16` underflow in the `eid-metrics` table's "Event" column width on terminals narrower than 55 columns: `terminal_width - 55` underflowed before the 45-character floor could apply, panicking in overflow-checked builds and wrapping to a huge value in release builds (leaving the column effectively uncapped). It now uses saturating subtraction. (#1897) (@YamatoSecurity)
-- Fixed `log-metrics` skewing the First/Last Timestamp columns for Splunk-JSON input: the timestamp fallback parsed the explicit UTC offset (e.g. `+09:00`) with `NaiveDateTime`, which discards the offset, storing the local wall-clock time as if it were UTC. The offset is now applied. (#1897) (@YamatoSecurity)
+- Fixed Splunk-JSON timestamps carrying an explicit UTC offset (e.g. `+09:00`) being parsed as a `NaiveDateTime`, which discards the offset and stored the local wall-clock time as if it were UTC — skewing the `log-metrics` First/Last Timestamp columns as well as the `eid-metrics`/`logon-summary` time ranges (`EventMetrics::stats_time_cnt` and `parse_evtx_datetime`). The timeline aggregators now share one offset-aware parser (`utils::parse_evtx_timestamp`) that applies the offset. (#1897) (@YamatoSecurity)
 - Fixed `-G` (GeoIP enrichment) with a custom rules config directory (`-c`) ignoring that directory's `geoip_field_mapping.yaml`: the lookup used the extensionless name `geoip_field_mapping`, which never matched the real file, so it always fell back to the default bundled mapping. (#1897) (@YamatoSecurity)
 - Fixed `count(field)` correlation rules producing a false-positive alert spanning longer than the rule's `timeframe`. In `judge_timeframe`'s sliding window, the branch handling a window that did not satisfy the count condition did an unchecked `add_data(right)` before dropping the left edge, pulling `records[right]` — already known to be outside the timeframe from the window's left edge — into the next window; when that record's field value was new it could push the distinct-value count over the threshold across a span longer than the timeframe. The slide now only drops the left edge and lets the inner loop re-extend `right` with the timeframe check, so no record beyond the timeframe is counted. Bare `count()` was unaffected. (#1896) (@YamatoSecurity)
 - Fixed narrow-terminal issues in the results summary: `_print_timeline_hist` did unchecked `usize` subtraction that could underflow on a very narrow terminal (huge allocation / hang) — it now skips the histogram when there is no room and uses `saturating_sub`; the rule-authors table computed its row count with a hard-coded `is_multiple_of(4)` instead of the actual column count (uneven tables at most widths) — now `authors_num.div_ceil(table_column_num)`; and the rule-authors table now falls back to a single column on very narrow terminals (≤72 chars) instead of overflowing/wrapping. Timeline output is unaffected. (#1858) (@YamatoSecurity)
